Fix compilers on Windows
The 19.16 version of the vs2017_win-64 package breaks the builds on
Windows. The reason is not totally clear, but is possibly related to how
the vcvars64.bat script is called, where the 19.16 version of the
package adds an argument to specify the toolset version. This ends up
causing SCons to be unable to link the library together after everything
builds. SCons throws a "file not found" error, even though the files are
clearly there. I suspect that there's a version mismatch somewhere with
the different MSVC parts and how Conda and SCons each set them up.

The resolution is to fix the vs2017_win-64 package to the 19.15 version.
This results in warnings about using the old method for specifying
compilers in conda-build, which (according to the message) will be
removed after conda-build 4.0. C'est la vie, we'll deal with it then.

Also, remove ruamel_yaml as a build-time dependency, due to changes in
the main Cantera branch.